[bucket] Expose bucket name in secrets.
The object storage controller creates secrets with bucket credentials
that have no reference to the parent BucketAccess object. Because of
this they cannot be linked to the managing app
(buckets.apps.cozystack.io) and are not displayed in the new dashboard.
This change patches the auxiliary helm release <bucket_name>-system to
include the bucket name in __its__ secret, so that the necessary secret
values is still presented to the user.

```release-note
[bucket] Expose bucket name in tenant secret.
```

Signed-off-by: Timofei Larkin <lllamnyp@gmail.com>